aufs: tiny, type-cast in a debug-print
To support both of 32/64 bit system.

+++ fs/aufs/rdu.c
@@ -278,9 +278,9 @@ static int au_rdu_verify(struct aufs_rdu *rdu)
 	    && rdu->verify[AufsCtlRduV_SZ_PTR] == sizeof(rdu))
 		return 0;
 
-	AuDbg("%u:%lu, %u:%lu\n",
-	      rdu->verify[AufsCtlRduV_SZ], sizeof(*rdu),
-	      rdu->verify[AufsCtlRduV_SZ_PTR], sizeof(rdu));
+	AuDbg("%u:%u, %u:%u\n",
+	      rdu->verify[AufsCtlRduV_SZ], (unsigned int)sizeof(*rdu),
+	      rdu->verify[AufsCtlRduV_SZ_PTR], (unsigned int)sizeof(rdu));
 	return -EINVAL;
 }
